requirements.txt
pandas
numpy
seaborn
python-dotenv
requests
google-cloud-bigquery
google-auth
google-api-core
gcloud
#-e .